Fix get_meas_interval()
The method was returning an array with only 1 element, the data for the `startDate`, instead of one element for each day of the entire requested interval. Inspecting the traffic, adding the `"type": 4` property seems to make it clear for the backend that we want to perform an interval operation and fixes the problem.